Acceptance: Get rid of `defer` in `EVS` tests
Replace `defer` with `quotas.Book...` in EVS tests

Simplify volume quota usage
