+++ opentelekomcloud/acceptance/evs/resource_opentelekomcloud_evs_volume_v3_test.go
@@ -4,11 +4,9 @@ import (
 	"fmt"
 	"regexp"
 	"testing"
-	"time"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
-	th "github.com/opentelekomcloud/gophertelekomcloud/testhelper"
 	"github.com/opentelekomcloud/terraform-provider-opentelekomcloud/opentelekomcloud/acceptance/common/quotas"
 
 	"github.com/opentelekomcloud/gophertelekomcloud/openstack/evs/v3/volumes"
@@ -23,9 +21,7 @@ const resourceVolumeV3Name = "opentelekomcloud_evs_volume_v3.volume_1"
 func TestAccEvsStorageV3Volume_basic(t *testing.T) {
 	var volume volumes.Volume
 	t.Parallel()
-	qts := []*quotas.ExpectedQuota{{Q: quotas.Volume, Count: 1}, {Q: quotas.VolumeSize, Count: 12}}
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, volumeQuotas(12))
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
@@ -52,9 +48,7 @@ func TestAccEvsStorageV3Volume_basic(t *testing.T) {
 
 func TestAccEvsStorageV3Volume_tags(t *testing.T) {
 	t.Parallel()
-	qts := []*quotas.ExpectedQuota{{Q: quotas.Volume, Count: 1}, {Q: quotas.VolumeSize, Count: 12}}
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, volumeQuotas(12))
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
@@ -80,9 +74,7 @@ func TestAccEvsStorageV3Volume_tags(t *testing.T) {
 func TestAccEvsStorageV3Volume_image(t *testing.T) {
 	var volume volumes.Volume
 	t.Parallel()
-	qts := []*quotas.ExpectedQuota{{Q: quotas.Volume, Count: 1}, {Q: quotas.VolumeSize, Count: 12}}
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, volumeQuotas(12))
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
@@ -103,9 +95,7 @@ func TestAccEvsStorageV3Volume_image(t *testing.T) {
 func TestAccEvsStorageV3Volume_timeout(t *testing.T) {
 	var volume volumes.Volume
 	t.Parallel()
-	qts := []*quotas.ExpectedQuota{{Q: quotas.Volume, Count: 1}, {Q: quotas.VolumeSize, Count: 12}}
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, volumeQuotas(12))
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
@@ -124,9 +114,7 @@ func TestAccEvsStorageV3Volume_timeout(t *testing.T) {
 
 func TestAccEvsStorageV3Volume_volumeType(t *testing.T) {
 	t.Parallel()
-	qts := []*quotas.ExpectedQuota{{Q: quotas.Volume, Count: 1}, {Q: quotas.VolumeSize, Count: 12}}
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, volumeQuotas(12))
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
@@ -146,9 +134,7 @@ func TestAccEvsStorageV3Volume_resize(t *testing.T) {
 	var volume volumes.Volume
 	var volumeUpScaled volumes.Volume
 	t.Parallel()
-	qts := []*quotas.ExpectedQuota{{Q: quotas.Volume, Count: 1}, {Q: quotas.VolumeSize, Count: 20}}
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, volumeQuotas(20))
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
